Fix video project drive choice 	Also fix a problem with size display
	* src/brasero-project-size.c
	(brasero_project_size_get_ruler_min_width),
	(brasero_project_size_get_media_string),
	(brasero_project_size_expose),
	(brasero_project_size_is_valid_drive),
	(brasero_project_size_build_menu),
	(brasero_project_size_scroll_event),
	(brasero_project_size_find_proper_drive),
	(brasero_project_size_set_context):
	* src/brasero-project-size.h:
	* src/brasero-project.c (brasero_project_switch):
	* src/brasero-utils.c (brasero_utils_get_time_string),
	(brasero_utils_get_time_string_from_size):
	* src/brasero-utils.h:
	* src/burn-basics.h:

svn path=/trunk/; revision=1010

+++ src/burn-basics.h
@@ -170,8 +170,8 @@ typedef enum {
 #define BRASERO_SIZE_TO_SECTORS(size, secsize)					\
 	(((size) / (secsize)) + (((size) % (secsize)) ? 1:0))
 #define BRASERO_BYTES_TO_DURATION(bytes)					\
-	(((bytes) * 1000000000) / (2352 * 75) + 				\
-	((((bytes) * 1000000000) % (2352 * 75)) ? 1:0))
+	(guint64) ((guint64) ((guint64) (bytes) * 1000000000) / (guint64) (2352 * 75) + 				\
+	(guint64) (((guint64) ((guint64) (bytes) * 1000000000) % (guint64) (2352 * 75)) ? 1:0))
 
 BraseroBurnResult
 brasero_burn_library_init (void);
